Reduced compilation time down to 35 seconds. 😊

+++ source/D3D11Helper/D3D11BufferHelper.cpp
@@ -221,7 +221,7 @@ SB_IDX CreateStructuredBuffer(const void* data, const size_t& size, const size_t
 	desc.BindFlags = D3D11_BIND_SHADER_RESOURCE;
 	desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
 	desc.MiscFlags = D3D11_RESOURCE_MISC_BUFFER_STRUCTURED;
-	desc.StructureByteStride = size;
+	desc.StructureByteStride = (UINT)size;
 
 	D3D11_SUBRESOURCE_DATA sbData;
 	sbData.pSysMem = data;
